requirements.txt
joblib==1.2.0
keyboard==0.13.5
numpy==1.21.6
opencv-python==4.7.0.72
pandas==1.3.5
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
